# ToneBlit performance and Method Implementation, Sprite Flashes, other fixes