The issues: ToneBlit is not very efficient, Battle tones and Gradual TintScreen are tinted differently, and Sprite flashes.

ToneBlit is very heavy performance-wise, which forces us to use two different ways to implement tintscreen in our screen:
1. We use ToneBlit for any individual tile, background and sprite when they're drawn.
2. After all the drawables are drawn, we use ToneBlit to the screen.

 - The 1st method, with a little modification, allows us to generate Sprite flashes that ignore the tintscreen (for example, a white flash with a dark-red tintscreen, it should appear completely white instead of red-ish). The 2nd doesn't allow that.
 - The 1st method is more efficient when the tintscreen remains static, but slower when the tone screen is gradually changing.
 - The 2nd is used for Gradual Tintscreen and Battle scenes. The 1st one is used anywhere else.

The ideal solution would be using the first to fix the flash bug. For that, we try to make the ToneBlit function more efficient. It sacrifices legibility in order to gain a little more of performance. These changes (cache variables, combining checks, etc.) improve the performance so there's not much difference for the 1st and 2nd method when a gradual tone screen happens, and therefore the 1st method is viable for everything.

To implement this:
 - Bitmap: We modify the ToneBlit to make it more efficient.
 - Sprite: We invert the positions of ToneBlit and BlendBlit. This way the flash color will have priority and not be tinted.
 - Spriteset_Map: Now in all the cases it uses the 1st method, so the code that allows the 2nd method is deleted.
 - Spriteset_Battle: We change from the 2nd method to the 1st one. We set the tone for the background and the battler sprites.
 - Background: We include the option of setting and getting a tone, and then draw it.
 - Screen: We eliminate the tone functions since no other class uses them.

Also, another fixes:
 - Game_Character flashes were limited to 128 no matter what the actual value was. An attribute flash_alpha has been added to account for this, and when the flash is set and invoked in its subclasses (Event, Vehicle, Player) we use that attribute.
 - Battle_Animation "flash_length" has been changed from 5 to 12, because the way it was before it was almost unnoticeable in comparison with RPG_RT.
 - Battle_Animation and Game_Interpreter_Map: The value of the flashes before with the << 3 operation meant the flash value for each part would be [0, 248]. With "x * 255 / 31", the values is really [0, 255].

Fix #1225