@Albeleon 2018-06-24
#1371 The original reason why I checked this:
"4.12: RM2000 (not tested in RM2003): Maybe it's my impression but when an attack animation shows up to an enemy, shouldn't the enemy flashes inside the animation (not afterwards) be clearer? In the video, sometimes they don't appear."