@Ghabry 2018-06-24
Could you split the single commit in two?
The bitmap.cpp change is completely unrelated to the rest.

The easiest way to do this hear would be doing "amend commit" and unchecking "bitmap.cpp". Then commiting "bitmap.cpp" using an additional commit.