@Ghabry 2018-06-24
For more context: The Tone/Saturation code had a optimisation that decided if the change is a gradient or static.
See #1017

But this change had visual issues e.g. for weather effects (can't find the issue number) and needs testing if @Albeleon fixed it.

In general the assumption that there is a "screen layer" and everything below it is affected by tone doesn't hold true anymore since Cherrys RPG2k3E release because each picture can be individually configured if it is affected by tone.

This PR needs some benchmarking.