@Ghabry 2018-06-25
This also fixes #1225