@Ghabry 2018-07-02
This also needs a rebase on upstrema to fix the Android build failure.