@Ghabry 2018-07-02
For benchmarking you can use this patch which unlocks the FPS but still keeps the updates at 60:

```diff
diff --git a/src/player.cpp b/src/player.cpp
index f646bc9c..d39cac16 100644
--- a/src/player.cpp
+++ b/src/player.cpp
@@ -248,7 +248,7 @@ void Player::Update(bool update_scene) {
 	static const double framerate_interval = 1000.0 / Graphics::GetDefaultFps();
 	next_frame = start_time + framerate_interval;
 
-#ifdef EMSCRIPTEN
+#if 1
 	// Ticks in emscripten are unreliable due to how the main loop works:
 	// This function is only called 60 times per second instead of theoretical
 	// 1000s of times.
@@ -303,12 +303,17 @@ void Player::Update(bool update_scene) {
 	}
 
 	Audio().Update();
-	Input::Update();
 
 	std::shared_ptr<Scene> old_instance = Scene::instance;
 
 	int speed_modifier = GetSpeedModifier();
 
+	double cur_time = (double)DisplayUi->GetTicks();
+	if (cur_time < next_frame) {
+		return;
+	}
+	Input::Update();
+
 	for (int i = 0; i < speed_modifier; ++i) {
 		Graphics::Update();
 		if (update_scene) {
```