@Ghabry 2018-07-02
With my benchmark I get the following FPS:

 Test  | Upstream  | This PR
---|---|---
No effect |  1050 | 1050
Tone gradient  | 650  | 970  
Saturation gradient | 623 | 970
Tone+Sat grad. | 500 | 960

The FPS measurement is not fully precise, error is ±50 FPS but you can see that this PR is significantly faster in all operations.